Add one test for helusers.oidc.ApiTokenAuthentication
This tests the happy path case, where the authentication passes. It's
just an example how this class could be tested.

The RSA keys found from keys.py are generated just for this purpose so
it's not a biggy that there are both the public and private keys in the
same file clearly readable.