HP-182 Use default key provider if one isn't given
If RequestJWTAuthentication doesn't receive a key provider in its
constructor it uses a default key provider that fetches keys from the
issuer server using the OIDC standard configuration URL.

This is how RequestJWTAuthentication is mainly meant to be used, using
the default key provider. The key provider constructor argument is
mainly for tests.