HP-182 Check API scope prefix if turned on in settings
This is for backwards compatibility.

If REQUIRE_API_SCOPE_FOR_AUTHENTICATION setting is True, a claim named
by API_AUTHORIZATION_FIELD setting is expected to be found and to be a
list of strings, one of which equals or starts with the string specified
by the API_SCOPE_PREFIX setting.

Similar functionality exists already in helusers.authz.UserAuthorization
class.

Added a reloading functionality into the
helusers.settings.api_token_auth_settings: the settings it contains gets
updated as settings are changed in tests.