Remove locked development requirements file
Python requirements depend on the Python version, operating system,
system architecture and what not. The requirements-dev.txt didn't
specify in any way in what kind of environment it was created. Now only
the base dependencies are mentioned in the requirements-dev.in file. The
developer is responsible for generating the requirements-dev.txt file
for their own environment.