HP-473 Start implementing a back channel logout endpoint
This commit adds a very basic view for the back channel logout URL, but
it doesn't do anything yet, other than respond with a 200 status.

The endpoint is excluded from the CSRF protection because the caller is
often another backend somewhere and they don't send any cookies in the
request.