Remove the key_provider argument from RequestJWTAuthentication
The key_provider argument was meant to support testing. But now the
tests provide proper stubs for authentication server so the tests don't
need to inject a key provider into RequestJWTAuthentication any more.
RequestJWTAuthentication can just use the real key provider which does
requests to the (stubbed in tests) authentication server.

The argument still IS there in the RequestJWTAuthentication __init__
method, but it just produces a warning if it's provided.