HP-473 Allow the callback to change the logout result
The user provided logout callback can report error situations that
terminate the logout handling and return the error result to the
original caller of the logout endpoint. An error is reported by
returning an HttpResponse with a status code between 400 and 599 from
the callback. Any other return value from the callback is ignored and
processing the logout event continues normally.