Validate aud claim in ApiTokenAuthentication
drf-oidc-auth changed from jwkest to authlib in a PR* where the accepted
audiences are read from a new setting and not from `get_audiences`
method. By default, every audience is accepted in authlib.

Changed claims_options to use our own audience setting.

The overriden claims_options method is used when drf-oidc-auth version
is 1.0.0 or greater. The old get_audiences will be used if the version
is less than 1.0.0.

* https://github.com/ByteInternet/drf-oidc-auth/pull/42