Fix drf-oidc-bug where the leeway used was way too big
In validate() call in oidc_auth.authentication.validate_claims the
leeway parameter was set as

`leeway=int(time.time()-api_settings.OIDC_LEEWAY)`

Which is wrong because the validation expects the leeway parameters to
be seconds. Not a point in time.

Leeway is 0 seconds by default in authlib.jose. Which is the same as in
jose which is used in RequestJWTAuthentication.