HP-1948 Change api token validation to support Keycloak tokens
In Tunnistamo the authorized scopes are in a claim like this:

    "https://example.com": [
      "scope",
      "scope2"
    ]

but in Keycloak the claims are for example:

    "authorization": {
      "permissions": [
        {
          "scopes": [
            "access"
          ]
        },
        {
          "scopes": [
            "second"
          ]
        }
      ]
    },

So the authorization field should have a way to tell where the scopes
are if they are deeper in the claims. The _authorized_api_scopes method
is therefore changed to support authorization field with field names
separated with dots.

In the previous example the scopes could be found with the following
authorization field: "authorization.permissions.scopes".

Additionally, the API_AUTHORIZATION_FIELD and API_SCOPE_PREFIX setting
usages are changed to support multiple values so that a project can
accept tokens from two different kind of issuers at the same time.