test: ad groups update of the user model
Test that the `update_ad_groups` function of the User-model works as
expected:

1. If the same group is mapped to multiple ad-groups and the ad-groups
input argument contains only some, the group is still persisted to user
instance since it was mapped to at least one of given ad-groups.

2. If the ad-groups input argument does not have any link to an existing
group of a user, the group is removed from the user instance.

3. If the ad-groups input argument contains links to new groups that the
user is not yet linked to, the group will be added to the user instance.